third_party/blink/web_tests/http/tests/inspector-protocol/cookies-protocol-test-expected.txt

Tests that cookies are set, updated and removed.
Test started
Enabling network

Running test: simpleCookieAdd
Setting Cookie
Success: true
Num of cookies 1
name: foo, value: bar1, domain: 127.0.0.1, path: /, session

Running test: simpleCookieChange
Setting Cookie
Success: true
Num of cookies 1
name: foo, value: second bar2, domain: 127.0.0.1, path: /, session

Running test: anotherSimpleCookieAdd
Setting Cookie
Success: true
Num of cookies 2
name: foo, value: second bar2, domain: 127.0.0.1, path: /, session
name: foo2, value: bar1, domain: 127.0.0.1, path: /, session

Running test: simpleCookieDelete
Deleting Cookie
Num of cookies 1
name: foo2, value: bar1, domain: 127.0.0.1, path: /, session

Running test: deleteAllCookies

Running test: sessionCookieAdd
Setting Cookie
Success: true
Num of cookies 1
name: foo, value: bar4, domain: 127.0.0.1, path: /, session

Running test: deleteAllCookies

Running test: nonSessionCookieZeroAdd
Setting Cookie
Success: true
Num of cookies 0

Running test: deleteAllCookies

Running test: nonSessionCookieAdd
Setting Cookie
Success: true
Num of cookies 1
name: foo, value: bar6, domain: 127.0.0.1, path: /, expires

Running test: deleteAllCookies

Running test: differentOriginCookieAdd
Setting Cookie
Success: true
Num of cookies 1
name: foo, value: bar7, domain: example.com, path: /, session

Running test: deleteAllCookies

Running test: invalidCookieAddDomain
Setting Cookie
Success: false
Num of cookies 0

Running test: deleteAllCookies

Running test: invalidCookieAddName
Setting Cookie
Success: false
Num of cookies 0

Running test: deleteAllCookies

Running test: secureCookieAdd
Setting Cookie
Success: true
Num of cookies 1
name: foo, value: bar, domain: 127.0.0.1, path: /, secure, session

Running test: deleteAllCookies

Running test: cookieAddHttpOnly
Setting Cookie
Success: true
Num of cookies 1
name: foo, value: bar, domain: 127.0.0.1, path: /, httpOnly, session

Running test: deleteAllCookies

Running test: cookieAddSameSiteLax
Setting Cookie
Success: true
Num of cookies 1
name: foo, value: bar, domain: 127.0.0.1, path: /, session, Lax

Running test: deleteAllCookies

Running test: cookieAddSameSiteLax
Setting Cookie
Success: true
Num of cookies 1
name: foo, value: bar, domain: 127.0.0.1, path: /, session, Strict

Running test: deleteAllCookies

Running test: setCookiesBasic
Adding multiple cookies
Num of cookies 8
name: cookie1, value: session, domain: localhost, path: /, session
name: cookie2, value: httpOnly, domain: localhost, path: /, httpOnly, session
name: cookie3, value: secure, domain: localhost, path: /, secure, session
name: cookie4, value: lax, domain: localhost, path: /, session, Lax
name: cookie5, value: expires, domain: localhost, path: /, expires
name: cookie6, value: .domain, domain: .chromium.org, path: /path, session
name: cookie7, value: domain, domain: www.chromium.org, path: /path, session
name: cookie8, value: url-based, domain: www.chromium.org, path: /, secure, session

Running test: deleteAllCookies

Running test: setCookiesWithInvalidCookie
Adding multiple cookies
Num of cookies 0

Running test: deleteAllCookies

Running test: deleteCookieByURL
Adding multiple cookies
Num of cookies 2
name: cookie1, value: .domain, domain: www.chromium.org, path: /, session
name: cookie2, value: .domain, domain: www.chromium.org, path: /, expires
Deleting Cookie
Num of cookies 1
name: cookie2, value: .domain, domain: www.chromium.org, path: /, expires

Running test: deleteAllCookies

Running test: deleteCookieByDomain
Adding multiple cookies
Num of cookies 2
name: cookie1, value: .domain, domain: .chromium.org, path: /path, session
name: cookie2, value: .domain, domain: .chromium.org, path: /path, expires
Deleting Cookie
Num of cookies 1
name: cookie2, value: .domain, domain: .chromium.org, path: /path, expires
Deleting Cookie
Num of cookies 0

Running test: deleteAllCookies

Running test: deleteCookieByDomainAndPath
Adding multiple cookies
Num of cookies 1
name: cookie1, value: .domain, domain: .chromium.org, path: /path, session
Deleting Cookie
Num of cookies 1
name: cookie1, value: .domain, domain: .chromium.org, path: /path, session
Deleting Cookie
Num of cookies 0

Running test: deleteAllCookies

Running test: nonUnicodeCookie
Adding multiple cookies
Num of cookies 1
name: cookie1, value: привет, domain: .chromium.org, path: /path, session

Running test: deleteAllCookies

Running test: setCookieViaFetch
Num of cookies 1
name: name, value: value, domain: 127.0.0.1, path: /inspector-protocol/network/resources, session

Running test: deleteAllCookies

Running test: printCookieViaFetch
Cookies as seen on server: "HTTP_COOKIE: foo=bar1\n"

Running test: deleteAllCookies
